Added in root privileges check to setup script

+++ Setup/15_10_dependencies.sh
@@ -1,25 +1,32 @@
 # Must be run as root
 
-apt-get update
-apt-get upgrade -y
+if [ "$(id -u)" != "0" ];
+then
+  echo "====================================================";
+  echo "Error - This script must be run with root privileges";
+  echo "====================================================";
+else
+  apt-get update
+  apt-get upgrade -y
 
-# Installs Apache
-apt-get install -y apache2
+  # Installs Apache
+  apt-get install -y apache2
 
-# Installs MySQL
-apt-get install -y mysql-server php5-mysql
+  # Installs MySQL
+  apt-get install -y mysql-server php5-mysql
 
-# Installs PHP
-apt-get install -y php5
+  # Installs PHP
+  apt-get install -y php5
 
-# Removes default index.html file
-rm -rf /var/www/html/index.html
+  # Removes default index.html file
+  rm -rf /var/www/html/index.html
 
-# Restarts Apache service
-service apache2 restart
+  # Restarts Apache service
+  service apache2 restart
 
-# Creates upload directory
-mkdir /var/www/html/uploads
+  # Creates upload directory
+  mkdir /var/www/html/uploads
 
-# Changes permissions so we can write to the uploads directory
-chown www-data:www-data /var/www/html/uploads
+  # Changes permissions so we can write to the uploads directory
+  chown www-data:www-data /var/www/html/uploads
+fi
